requirements.txt
Flask==2.1.1
Flask_RESTful==0.3.9
mongoengine==0.24.1
PyJWT==2.3.0
python-dotenv==0.20.0
redis~=4.3.0
celery~=5.2.6